LayoutTests/svg/animations/animate-setcurrenttime-expected.txt

SVG 1.1 dynamic animation tests

 PASS plain.x.animVal.value is 0
PASS sequential.x.animVal.value is 0
PASS accumulating.x.animVal.value is 0
PASS repeating.x.animVal.value is 0
PASS nested.x.animVal.value is 0
PASS plain.x.animVal.value is 0
PASS sequential.x.animVal.value is 0
PASS accumulating.x.animVal.value is 0
PASS repeating.x.animVal.value is 0
PASS nested.x.animVal.value is 0
PASS plain.x.animVal.value is 128
PASS sequential.x.animVal.value is 128
PASS accumulating.x.animVal.value is 128
PASS repeating.x.animVal.value is 128
PASS nested.x.animVal.value is 128
PASS plain.x.animVal.value is 0
PASS sequential.x.animVal.value is 0
PASS accumulating.x.animVal.value is 0
PASS repeating.x.animVal.value is 0
PASS nested.x.animVal.value is 0
PASS plain.x.animVal.value is 32
PASS sequential.x.animVal.value is 32
PASS accumulating.x.animVal.value is 128
PASS repeating.x.animVal.value is 128
PASS nested.x.animVal.value is 32
PASS plain.x.animVal.value is 64
PASS sequential.x.animVal.value is 32
PASS accumulating.x.animVal.value is 256
PASS repeating.x.animVal.value is 0
PASS nested.x.animVal.value is 64
PASS plain.x.animVal.value is 96
PASS sequential.x.animVal.value is 64
PASS accumulating.x.animVal.value is 192
PASS repeating.x.animVal.value is 128
PASS nested.x.animVal.value is 96
PASS plain.x.animVal.value is 128
PASS sequential.x.animVal.value is 128
PASS accumulating.x.animVal.value is 128
PASS repeating.x.animVal.value is 128
PASS nested.x.animVal.value is 128
PASS plain.x.animVal.value is 0
PASS sequential.x.animVal.value is 0
PASS accumulating.x.animVal.value is 0
PASS repeating.x.animVal.value is 0
PASS nested.x.animVal.value is 0
PASS plain.x.animVal.value is 0
PASS sequential.x.animVal.value is 0
PASS accumulating.x.animVal.value is 0
PASS repeating.x.animVal.value is 0
PASS nested.x.animVal.value is 32
PASS plain.x.animVal.value is 0
PASS sequential.x.animVal.value is 0
PASS accumulating.x.animVal.value is 0
PASS repeating.x.animVal.value is 0
PASS nested.x.animVal.value is 64
PASS plain.x.animVal.value is 0
PASS sequential.x.animVal.value is 0
PASS accumulating.x.animVal.value is 0
PASS repeating.x.animVal.value is 0
PASS nested.x.animVal.value is 96
PASS plain.x.animVal.value is 0
PASS sequential.x.animVal.value is 0
PASS accumulating.x.animVal.value is 0
PASS repeating.x.animVal.value is 0
PASS nested.x.animVal.value is 128
PASS plain.x.animVal.value is 0
PASS sequential.x.animVal.value is 0
PASS accumulating.x.animVal.value is 0
PASS repeating.x.animVal.value is 0
PASS nested.x.animVal.value is 0
PASS plain.x.animVal.value is 32
PASS sequential.x.animVal.value is 32
PASS accumulating.x.animVal.value is 128
PASS repeating.x.animVal.value is 128
PASS nested.x.animVal.value is 0
PASS plain.x.animVal.value is 64
PASS sequential.x.animVal.value is 32
PASS accumulating.x.animVal.value is 256
PASS repeating.x.animVal.value is 0
PASS nested.x.animVal.value is 0
PASS plain.x.animVal.value is 96
PASS sequential.x.animVal.value is 64
PASS accumulating.x.animVal.value is 192
PASS repeating.x.animVal.value is 128
PASS nested.x.animVal.value is 0
PASS plain.x.animVal.value is 128
PASS sequential.x.animVal.value is 128
PASS accumulating.x.animVal.value is 128
PASS repeating.x.animVal.value is 128
PASS nested.x.animVal.value is 0
